requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
yfinance==0.2.40
ta==0.11.0
anthropic==0.86.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
chromadb==0.6.3
sentence-transformers==3.4.1
